// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ACCOUTANT: 'ACCOUTANT',
  ADMIN: 'ADMIN',
  STUDENT: 'STUDENT'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const StudentStatus = {
  NOT_REGISTERED: 'NOT_REGISTERED',
  REGISTERED: 'REGISTERED'
} as const

export type StudentStatus = (typeof StudentStatus)[keyof typeof StudentStatus]


export const InvoiceType = {
  SCHOOL_FEES: 'SCHOOL_FEES',
  OTHER_FEES: 'OTHER_FEES'
} as const

export type InvoiceType = (typeof InvoiceType)[keyof typeof InvoiceType]


export const InvoiceStatus = {
  PENDING: 'PENDING',
  PARTIAL: 'PARTIAL',
  PAID: 'PAID'
} as const

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus]


export const PaymentMethod = {
  BANK: 'BANK',
  MOBILE_MONEY: 'MOBILE_MONEY',
  CARD: 'CARD'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]